updated map changes

+++ iclub-model/src/main/java/za/co/iclub/pss/ws/model/IclubPropertyModel.java
@@ -20,8 +20,8 @@ public class IclubPropertyModel {
 	private Long iclubRoofType;
 	private String PRegNum;
 	private String PAddress;
-	private Long PLat;
-	private Long PLong;
+	private Double PLat;
+	private Double PLong;
 	private Integer PPostalCd;
 	private Integer PNoclaimYrs;
 	private String PRentFurYn;
@@ -108,16 +108,16 @@ public class IclubPropertyModel {
 	public void setPAddress(String pAddress) {
 		PAddress = pAddress;
 	}
-	public Long getPLat() {
+	public Double getPLat() {
 		return PLat;
 	}
-	public void setPLat(Long pLat) {
+	public void setPLat(Double pLat) {
 		PLat = pLat;
 	}
-	public Long getPLong() {
+	public Double getPLong() {
 		return PLong;
 	}
-	public void setPLong(Long pLong) {
+	public void setPLong(Double pLong) {
 		PLong = pLong;
 	}
 	public Integer getPPostalCd() {
